clean code in Item, Item balancing, minor bugfix, Character movement speed gets calculated (#158)
* put method logic in smaller methods in order to make them less complex

* removed not used using directive

* Clean code in Item, bugfix in Characters.cs

made variables in item private and added properties
added maxPsychoLevel variable

* removed not used using directives, implemented changes of Item in ShopManager

* changed Item scriptable Objects, deleted empty directories

* character movementSpeed as impact on actual movementSpeed now, bugfix in GetPercentage method in Weapon

+++ WASD/Assets/Scripts/Weapons/Weapon.cs
@@ -57,6 +57,6 @@ public abstract class Weapon : MonoBehaviour, IBuyable
 
     private float GetPercentage(int value)
     {
-        return (value + 100) / 100;
+        return ((float) value + 100) / 100;
     }
 }
